backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pytest
httpx
openai
python-dotenv
faster-whisper

